Fix bug in Lv3 profit calculation
The rise in profit from a berry bush should be linear as its
neighbouring contributors rise from 0 to 7, after which the profit is
capped. The formula specified for Level 3 in Issue #41 had an extra
divide by six, greatly reducing the value of bushes with less than 8
contributors. This left a sudden spike in value at 8 contributors.

This bug only affected bushes that did not have maxed-out contributor
counts. Because very easy to max-out contributor counts, the overall
impact of the bug on the game's strategy was limited.