Rework mouse events for performance
Switching to a single listener on the stage object rather than on
each sprite saves EaselJS from having to do collision tests for
each mouse event. Our calculation to find the cell is much faster
because we assume the cells are in a grid.

Switching to a single listener also allows us to leverage pressmove
events. These are like mouseover events, but are only triggered
by dragging after mousedown. That saves a lot of CPU when you're not
doing anything.

No JS executes unless the user is actually interacting with the game,
fixing Issue #121.