Add minimum multiplier of 0.1 for Lv1 Profit
When there are no trees, blueberry bushes provide $0. This may be
confusing for new players as their actions don't seem to have any
effect.

The minimum multiplier is small to ensure that adding the first tree
will always result in a profit increase. Though, perhaps alternative
profit functions should be considered to keep the value of tree-less
blueberry bushes identical across levels 1, 2 and 3.