Fix or suppress W3 Validator errors
Error: An img element must have an alt attribute, except under certain
conditions.
Fix: The logo doesn't need alt-text. It's redundant due to the title,
so we can just suppress it.

Error: Element h4 not allowed as child of element span in this context.
Fix: Replace h4 with a span and tweak the margins to match.